Lab6/requirements.txt
fastapi 
pydantic
joblib
pandas
uvicorn
scikit-learn
PyYAML
ucimlrepo
